Fucking garbage (minor refactoring)
