+++ lizardwizards/src/main/java/com/lizardwizards/lizardwizards/client/LizardWizardsApplication.java
@@ -1,8 +1,8 @@
-package com.lizardwizards.lizardwizards;
+package com.lizardwizards.lizardwizards.client;
 
+import com.lizardwizards.lizardwizards.core.Utils;
 import javafx.application.Application;
 import javafx.stage.Stage;
-import javafx.stage.StageStyle;
 
 import java.io.IOException;
 
@@ -10,7 +10,7 @@ public class LizardWizardsApplication extends Application {
     @Override
     public void start(Stage stage) throws IOException {
         stage.setOnCloseRequest(e -> {e.consume(); Utils.closeProgram(stage);});
-        stage.setTitle("LizardWizards");
+        stage.setTitle("Lizard Wizards");
         Utils.changeScene(stage, 0);
         stage.show();
     }
