+++ lizardwizards/src/main/java/com/lizardwizards/lizardwizards/core/EntitySprite.java
@@ -1,4 +1,4 @@
-package com.lizardwizards.lizardwizards;
+package com.lizardwizards.lizardwizards.core;
 
 import javafx.scene.shape.Rectangle;
