+++ lizardwizards/src/main/java/com/lizardwizards/lizardwizards/core/PlayerControls.java
@@ -1,4 +1,4 @@
-package com.lizardwizards.lizardwizards;
+package com.lizardwizards.lizardwizards.core;
 
 import javafx.scene.Scene;
 import javafx.scene.input.KeyCode;
