+++ lizardwizards/src/main/java/com/lizardwizards/lizardwizards/core/Weapon.java
@@ -1,4 +1,4 @@
-package com.lizardwizards.lizardwizards;
+package com.lizardwizards.lizardwizards.core;
 
 import java.util.List;
