+++ lizardwizards/src/main/java/module-info.java
@@ -4,6 +4,8 @@ module com.lizardwizards.lizardwizards {
 
     requires org.controlsfx.controls;
 
-    opens com.lizardwizards.lizardwizards to javafx.fxml;
-    exports com.lizardwizards.lizardwizards;
+    exports com.lizardwizards.lizardwizards.core;
+    opens com.lizardwizards.lizardwizards.core to javafx.fxml;
+    exports com.lizardwizards.lizardwizards.client;
+    opens com.lizardwizards.lizardwizards.client to javafx.fxml;
 }
