+++ lizardwizards/src/main/resources/com/lizardwizards/lizardwizards/game-view.fxml
@@ -8,7 +8,7 @@
 
 <AnchorPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="com.lizardwizards.lizardwizards.GameController"
+            fx:controller="com.lizardwizards.lizardwizards.core.GameController"
             prefHeight="400.0" prefWidth="600.0">
     <Label text=":)"/>
 </AnchorPane>
