+++ lizardwizards/src/main/resources/com/lizardwizards/lizardwizards/menu-view.fxml
@@ -6,7 +6,7 @@
 
 <?import javafx.scene.control.Button?>
 <VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="com.lizardwizards.lizardwizards.MenuController">
+      fx:controller="com.lizardwizards.lizardwizards.client.MenuController">
     <padding>
         <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
     </padding>
